Rename rangeMatchModifiers to rangeContainsModifiers
