+++ src/utils/index.ts
@@ -2,5 +2,5 @@ export * from "./addToRange.js";
 export * from "./dateMatchModifiers.js";
 export * from "./rangeContainsDayOfWeek.js";
 export * from "./rangeIncludesDate.js";
-export * from "./rangeMatchModifiers.js";
+export * from "./rangeContainsModifiers.js";
 export * from "./typeguards.js";
