+++ src/utils/rangeContainsModifiers.ts
